+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Features**:
 
 - Add metadata support for the `/upload` endpoint. ([#6028](https://github.com/getsentry/relay/pull/6028))
+- Infer user agents and client addresses in the V2 standalone span pipeline. ([#6047](https://github.com/getsentry/relay/pull/6047))
 
 **Bug Fixes**: